Updated/Password-cracker/requirements.txt
# Web Framework
Flask==3.0.2
Werkzeug==3.0.1
Flask-WTF==1.2.1

# Security
bcrypt==4.1.2
cryptography==42.0.2
PyJWT==2.8.0
python-dotenv==1.0.1

# Production Server
gunicorn==21.2.0
gevent==24.2.1

# Testing
pytest==8.0.0
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# Monitoring
sentry-sdk==1.39.1
prometheus-client==0.19.0

# Utilities
python-dateutil==2.8.2
requests==2.31.0
urllib3==2.2.0
certifi==2024.2.2 